Update backend models and controllers
