+++ backend/models/Cart.js
@@ -1,5 +1,5 @@
 // Define Cart schema
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 
 const cartSchema = new mongoose.Schema({
   userId: {
@@ -18,4 +18,4 @@ const cartSchema = new mongoose.Schema({
 
 const Cart = mongoose.model('Cart', cartSchema);
 
-module.exports = Cart;
+export default Cart;
