+++ backend/util.js
@@ -1,0 +1,16 @@
+import jwt from 'jsonwebtoken';
+
+// Middleware to verify JWT token
+function verifyToken(req, res, next) {
+    const token = req.cookies.token;
+    if (!token) return res.status(401).send('Unauthorized');
+
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
+        if (err) return res.status(403).send('Invalid token');
+        req.user = decoded;
+        next();
+    });
+}
+
+
+export { verifyToken };
